Renamed libraries to src and fixed BYJStepper::step function
