+++ BasicOperationPlus/src/Direction/Direction.cpp
